requirements.txt
tensorflow==2.11.0
Pillow>=9.1.1
pytz>=2022.6
tqdm>=4.64.1
tf-models-official==2.11.3
scikit-learn>=1.2.1
albumentations>=1.3.0